ESPP-17 Move exception class from Elasticsuite\Index\Exception to Elasticsuite\Exception
