+++ api/src/Elasticsuite/Standard/src/Index/MutationResolver/CreateIndexMutation.php
@@ -19,7 +19,7 @@ namespace Elasticsuite\Index\MutationResolver;
 use ApiPlatform\Core\Exception\InvalidArgumentException;
 use ApiPlatform\Core\GraphQl\Resolver\MutationResolverInterface;
 use Elasticsuite\Catalog\Repository\LocalizedCatalogRepository;
-use Elasticsuite\Index\Exception\LogicException;
+use Elasticsuite\Exception\LogicException;
 use Elasticsuite\Index\Service\IndexOperation;
 use Elasticsuite\Metadata\Repository\MetadataRepository;
